pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "A Python IoT/Hardware application template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bleak>=0.22",
    "httpx>=0.27",
    "paho-mqtt>=2.1",
    "pydantic-settings>=2.7",
    "pyserial>=3.5",
    "python-dotenv>=1.0",
    "structlog>=24.4",
    "websockets>=14.0",
]

[project.optional-dependencies]
rpi = ["gpiozero>=2.0", "smbus2>=0.5"]

[project.scripts]
myapp = "myapp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myapp"]

# ── Ruff ──────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
src = ["src", "tests"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["myapp"]

# ── mypy ──────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
packages = ["myapp"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "serial.*",
    "gpiozero.*",
    "smbus2.*",
    "paho.*",
    "bleak.*",
]
ignore_missing_imports = true

# ── pytest ────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]
